# pix-app/src/main/resources/static/openapi.yaml
openapi: 3.0.3

info:
  title: mini-pix
  description: |
    Kafka laboratory simulating the PIX payment system following **BACEN** and **ISO 20022** specifications.

    ## Authentication
    All endpoints require a JWT bearer token in the `Authorization` header.
    The sender's document (CPF/CNPJ) is extracted from the `sub` claim — it is never sent in the request body.

    ## Test JWT
    **Secret:** `minipix-secret-key-for-local-dev` | **Algorithm:** HS256 | **sub:** `52998224725`
    ```
    Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI1Mjk5ODIyNDcyNSIsIm5hbWUiOiJFZHVhcmRvIEZvbnRlbmVsZSIsImlhdCI6MTUxNjIzOTAyMn0.ZEY0hTMxUHyjLhTet_sld-64t-mIET4brsfbDLu522o
    ```
  version: 0.0.1-SNAPSHOT

servers:
  - url: http://localhost:8080
    description: Local

security:
  - bearerAuth: []

tags:
  - name: Transactions
    description: Pix transaction lifecycle

paths:
  /pix/transactions/resolve:
    post:
      tags:
        - Transactions
      summary: Resolve receiver information
      description: |
        Pre-confirmation step. Resolves a Pix key and returns the receiver's data for the payer
        to review before confirming the transfer. Follows the **BACEN DICT specification** and is
        required by regulation before any funds are transferred.

        The following checks run **in parallel** using virtual threads:
        - **DICT** — resolves the Pix key to account, receiver, and institution data
        - **Anti-fraud** — checks whether the key or the sender is flagged
        - **Limits** — verifies whether the sender has sufficient limit for the requested amount

        All parallel calls are subject to a **200ms timeout** aligned with BACEN's SLA for DICT responses.
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/DeviceId'
        - $ref: '#/components/parameters/ForwardedFor'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PixResolveRequest'
            example:
              receiverKey: joao.silva@gmail.com
              amount: 250.00
      responses:
        '200':
          description: Receiver information resolved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PixResolveResponse'
        '422':
          description: Transaction not allowed — fraud risk detected or insufficient limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '503':
          description: External service unavailable or timed out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

  /pix/transactions/init:
    post:
      tags:
        - Transactions
      summary: Initiate a Pix transfer
      description: |
        Initiates a Pix transfer after payer confirmation.

        This endpoint does not process the transaction synchronously. It validates the request,
        checks idempotency, and publishes the transaction to Kafka for asynchronous processing
        by the downstream modules (validator, auditor, notificador). The final outcome is
        delivered via notification — not in this response.

        Internally follows **ISO 20022 pacs.008** (FIToFICustomerCreditTransfer) for the Kafka message structure.
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PixInitRequest'
            example:
              receiverKey: joao.silva@gmail.com
              amount: 250.00
              description: Reembolso do almoço
      responses:
        '202':
          description: Transaction accepted for processing
        '422':
          description: Duplicate transaction — idempotency key already used
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    CorrelationId:
      name: correlation-id
      in: header
      required: true
      description: UUID for distributed tracing across all internal services
      schema:
        type: string
        format: uuid
      example: f47ac10b-58cc-4372-a567-0e02b2c3d479

    IdempotencyKey:
      name: idempotency-key
      in: header
      required: true
      description: UUID generated by the client to prevent duplicate transactions
      schema:
        type: string
        format: uuid
      example: 3d3b5a1e-1c2d-4e5f-8a9b-0c1d2e3f4a5b

    DeviceId:
      name: device-id
      in: header
      required: true
      description: Device fingerprint consumed by the anti-fraud service
      schema:
        type: string
      example: device-test-001

    ForwardedFor:
      name: forwarded-for
      in: header
      required: true
      description: Original client IP consumed by the anti-fraud service
      schema:
        type: string
      example: 189.120.45.32

  schemas:
    PixResolveRequest:
      type: object
      required:
        - receiverKey
        - amount
      properties:
        receiverKey:
          type: string
          description: The receiver's Pix key (CPF, CNPJ, email, phone, or random key)
        amount:
          type: number
          format: decimal
          description: Transfer amount in BRL

    PixInitRequest:
      type: object
      required:
        - receiverKey
        - amount
      properties:
        receiverKey:
          type: string
          description: The receiver's Pix key resolved in the previous step
        amount:
          type: number
          format: decimal
          description: Transfer amount in BRL confirmed by the payer
        description:
          type: string
          description: Optional message to the receiver

    PixResolveResponse:
      type: object
      properties:
        receiver:
          $ref: '#/components/schemas/ReceiverInfo'
        account:
          $ref: '#/components/schemas/AccountInfo'
        institution:
          $ref: '#/components/schemas/InstitutionInfo'
        key:
          type: string
        keyType:
          $ref: '#/components/schemas/PixKeyType'
        keyRegisteredAt:
          type: string
          format: date-time

    ReceiverInfo:
      type: object
      properties:
        name:
          type: string
        personType:
          $ref: '#/components/schemas/PersonType'
        maskedDocument:
          type: string
          example: "***.***.987-**"
        document:
          type: string
          example: "12345678987"

    AccountInfo:
      type: object
      properties:
        maskedBranch:
          type: string
          example: "0001"
        maskedAccountNumber:
          type: string
          example: "****-9"
        accountType:
          $ref: '#/components/schemas/AccountType'

    InstitutionInfo:
      type: object
      properties:
        name:
          type: string
          example: Nubank
        ispb:
          type: string
          example: "18236120"
        compe:
          type: string
          example: "260"

    ProblemDetail:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string

    PixKeyType:
      type: string
      enum: [CPF, CNPJ, EMAIL, TELEFONE, CHAVE_ALEATORIA]

    AccountType:
      type: string
      enum: [CACC, SVGS, TRAN]

    PersonType:
      type: string
      enum: [PESSOA_FISICA, PESSOA_JURIDICA]